Keep moving the concept of an array value in a field deeper into the Fields and Field classes.  :group_request is now of conversion type :group as is :list_request.  These take an array as their singular value.  As such, they needed a special case in the set_value method of Field.
But, we can now set :text_lines to an array of lines and
:group_request to an array of symbols: the later being stored as a
sequence of shorts as a single value.

Also rearranged how Fields get put into a Request.

Improved the singularize and pluralize concepts and implemented
several methods in String as well as Symbol to help working with
singular and plural values.

Finally, we have CT from the queue status page working as well as
addtxt from the show call page.


git-svn-id: svn://p51.austin.ibm.com/raptor/trunk@165 c4f43c74-3acd-11dc-88bc-9584e4a9cacb