Fixed editable fields.
I removed the non-standard attributes from the spans that surround the
in place editors.  These attributes use to specify paramters to
customize the editors.

Instead, I added a help method called add_page_setting.  This helper
takes a name and an object and adds it as a hash entry to the
page_settings attribute.  I modified the qs layout to dump the
contents of page_settings out as a javascript object calling it
pageSettings and each hash entry becomes an attibute with the value
that was passed.

The Javascript on load routine now finds the customization paramters
needed for each in place editor from pageSettings which is keyed off
of the id of the element.

I also changed how owner, resolver, and next queue on the queue status
page was rendered moving towards using partials.  Currently it is
still a mix of Ruby code and partials.  But the short term benefit is
that when the field is changed via the in placed editor, it is now
rendered the same.


git-svn-id: svn://p51.austin.ibm.com/raptor/trunk@257 c4f43c74-3acd-11dc-88bc-9584e4a9cacb