Added a "Queue back to ..." button
The button appears if the owner or resolver of the call is not the
same as the queue the call is currently on.  If picked by the user,
the owner's or resolver's person queue is picked, "requeue" is picked
as the action, the ct box is unchecked, the add time box is unchecked,
and the text box is cleared.  The main purpose is to forward a PMR
back to the owner or resolver.  The resolver takes precedence over the
owner.