Another small step
Start to implement the queue's calls table with a somewhat empty html
page that produces a call back to get the list of calls as a json
array.  Each element in the array will also make a call back to get
the data needed to populate that row of the table.  The concept is
that the initial render will be quick and the individual calls will be
filled in with data as the server respondes.  I'm hoping that the
individual calls will get cached up locally (eventually).