Removed final traces of full refresh, fixed caching problem
Modified qs and psar controller to use the @no_cache attribute rather
than the full refresh attribute.

The fragment caching had a problem.  The host is part of the fragment
cache's key.  If the host has an alias, that presents a problem.
e.g. tcp237 and tcp237.austin.ibm.com cause two versions of the cache
to be created but only one version gets invalidataed when the database
record changes.  The code now sets the host field to the host without
the domain.  That can cause problems but all the solutions I can think
of can cause a problem too.