Improved pmrs_controller and error handling
Raptor can now look at a PMR that is not on a queue.  If the pmr has a
view, the pmrs_controller redirects to the call to display the call.
Otherwise, it displays the pmr with its own view.  Some duplicate code
here that needs to be cleaned up sometime.

Also got off on a tangent catching errors and making them better.
This lead to the discovery that the from_param! were not throwing
exceptions like they were suppose to.