Final touchup for center
Some minor touchups for displaying centers.  Somehow, often, the
@exception can not be accessed from the error views so a the instance
variable is fected from "controller".  Also, center "000" is in the
database but it does not really exist.  It is returned by retain for
"none".  Added special code in the load of the center to not really
fetch the center if it was "000".