Fixed some of the todos:
Verify EPCAAT and flag if not complete.  Probably color the
&ldquo;link-etc&rdquo; area red in the QS page and the EPCAAT
button on the call page red.

The PMR itself needs a title attribute to indicate why it is
colored the way it is.  I'm not sure where to put this.

The darkened color of a PMR is not correct.  It should key off
the owner of the queue the call is on, if any.  Not the
owner of the call itself.

I don't like the new title used for calls.