Removed not null restriction from customer id in pmrs.
I'm not 100% sure how to do this.  The DB guys hate nulls.  I do too.
I guess I should make it default to "not found yet" or something to
that effect so that we avoid the three level logic problems with
nulls.  But there is a general problem that needs to be address so
I'll add that to my todo list and not address it right now.