Fixes to make Raptor work with Ruby 1.8 and 1.9
Added an initializer that gives Fixnum and ord function for 1.8 if it
does not already exist.  This allows "asd"[0].ord to always work.

Tried via egrep to find all occurrences of a string with a single
index (like "abcd"[0]) and added .ord.  Also added .chr to convert
back.

Had to fix the code in qs_helpers that defines the td, scan, etc
methods because Array#to_s changed to add in commas, etc.  i.e. in
1.9, [ "abc", "def" ].to_s => "[\"abc\", \"def\"]"