Hooked up PUT to bump counter
Decorated the already selected values at the top of the select list
with Aqua so they standout.  I am considering making the table more
elaborate with groupings.  Given that this is a table and not a select
element with options, I could do a multi-level menu.  But I think I'll
leave it like it is and see what type of feedback I get.