Unfinished redo of OPC, QS page, and combined_pmrs page
Previously OPC and Executive Summary could only by done on the call
page and not from the QS page nor from a closed PMR.  This is the
second commit to redo the QS page and tuck the smarts of how to do the
OPC update down into the OPC model (which isn't exactly the right
place for it).  Currently OPC on the call page is working again with
the QS page rearranged but not working and the PMR page also started
on the rearangement.

Conflicts:
	app/controllers/retain/call_controller.rb
	app/controllers/retain/pmrs_controller.rb
	app/helpers/retain/call_helper.rb
	app/helpers/retain/call_update_helper.rb
	app/views/shared/retain/_opc_form.html.erb
	config/routes.rb