Completed the revamp of OPC interface
The QS page now has a selector to pick what to do which can be to update
the call, update the OPC entry, do a CT, and dispatch or undispatch.
Added some visual clues so that the user can see if OPC has been done
and added a text span in the update form of the last service given.

Conflicts:
	app/controllers/retain/call_controller.rb
	app/views/retain/call/show.html.erb
	app/views/shared/retain/_call_update.html.erb
	public/javascripts/retain/call.js