Made creation_date in combined_pmr not null.
I removed the redirect from a PMR to the primary call while debugging.
I added it back in.  While I was at it, I modified how creation_date
works.  A PMR is now required to have a non-null creation date.  The
call to valid? finds the creation_date and the code that creates the new
database entries uses that to initialize the field.  find_existing_pmr
was changed as well.  We'll see how all this goes...