Changed scan_pmrs to scan only 1000 PMRs
The script dies if there are too many probably due to a memory leak.
1000 seems to work all the time.  It should be that a PMR does not
change for at least a month after it has been closed.  So, as long as
there are 28,000 PMRs or less, each PMR should get updated at least one
after it has been closed and before it is deleted from Retain.

Note that the load of Combined::Pmr sets updated_at to the current time
on all of the normal, non-error paths.