Added log of retain calls and fixed no PMR issue
If a queue has a call (usually a backup) for a PMR that has been deleted
from Retain, Raptor use to throw an error.  That was solved I hope.

Also, the sendit message was added to the logger at an info level hoping
to give me a clue why Raptor is making so many retain calls.