Fixed (I hope) conversion bug
Fairly often, line 14 in _body.html.erb would throw an error because of
a mismatch in the encoding saying that UTF-8 was not compatible with
ISO-8859-1.  I added a debug statement and each div to print out the
encoding and some were ASCII, some were ISO08859-1, and some were
UTF-8.  That surprised me because I expected all of the to be UTF-8.  So
I added an encode statement to make all of them UTF-8 and we'll see if
that fixes the error.