Changes to fix various issues
Trying to track down problems with encoding and finally discovered that
it was caused by content_for.  It ends up making everything encoded as
US-ASCII.  That fix will be not here but in the process I tried changing
'coding: utf-8' to 'encoding: utf-8' for all of the views.  It had not
good effect but I left them.

The call layout needed the opc javascript.

Changes the raptor script which is put in /etc/rc.d/init.d to use bundle
exec rather than prvm-wrapper.