Fixed bug causing uq_cached_pmrs_triple to trigger
If a call is on a queue but the PMR for the call has been deleted from
Retain, we were trying to find a PMR with the same problem, branch, and
country with deleted set to false and the creation_date set to the bogus
value of 01/01/2014.  Since the PMR had already been deleted, the
process that updates the PMRs would mark the PMR as deleted on its next
run.  After that, viewing the same queue would cause the constraint to
fire.  The PMR was not in retain so the normal create_from_options would
fail (in find_existing_pmr).  The find_or_new would first look for a PMR
that was not deleted -- an not find it.  Then it would try to create one
but the key does not include the deleted bit so the create would fail
due to a duplicate key (same four-tuple).

The fix was to do the find without specifying the deleted attribute
since the record may or may not be deleted.

Other fix is to the owner and resolver of the view when the PMR has been
deleted.