Changed requeue from team to personal
In the past, the resolver was always set and the owner was also set if
the PMR was a US PMR.  That was changed to never set the owner but the
logic of the resolve is the same.  For any PMR, when queued from a team
queue to a personal queue, the resolve is set to the owner of the queue
the call is requeued to.