The list of retain nodes is now stored in the db
This is not a new change.  While moving tcp237 from one lab to another I
discovered that the old method of specifying how to get to retain was
never deleted at least from this tree.  retain.yml is not used and
Retain::Config is not used.

Rather, in the db, set up by hand, is a list of nodes and node selectors
and a retuser picks which node select which picks which node.  A flag
called TUNNELLED can be set at load type by creating a host specific
file as was being done for tcp237 (and may need to be recreated but
currently it is able to access Retain directly).