Added ECPAAT lines to update dialog
If Environment, Customer Rep, or Testcase are not present in ECPAAT,
then empty header lines are added to the update dialog.

For Problem, if it is not present, it is added as well.  But if it is
present and more than 2 weeks old, then the heading as well as the
previous entry is added hoping to encourage the user to update it.
We'll see how this flies.