Fix some QM nits
The list of users for a rotation group was limited to just the active
users.  This prevented making an inactive user active.

The QM page restricted users from outside the group from viewing the
page.  This was removed with a comment that there might be advantages
to keeping it that way.