Fixed newtext
Added a call to translit to change the funky UTF 8 characters into
normal ASCII characters.  e.g. when I put a single quote in my email,
then copy and paste it into a retain update, it would get scampled up
because the single quote was being entered as \xE2\x80\x98.  The call
to translit fixes it back to just a simple ASCII single quote.

This maybe should be done other places.  I tried adding it to
retain/field.rb but it borked because by that point, I have already
munged the string.